+++ tests/testthat/test-common-graphs.R
@@ -1,4 +1,3 @@
-context("Graphics")
 test_that("Basefont selection works", {
   dat <- get_base_fontfamily()
   expect_type(dat, "character")
@@ -26,9 +25,30 @@ test_that("Domain VaR heatmap", {
   expect_s3_class(gg, "gg")
 })
 
-test_that("Scatterplot", {
+test_that("generate_scatterplot() throws a warning", {
   data(mc_simulation_results)
-  gg <- generate_scatterplot(mc_simulation_results, scenario_id = 50)
+  expect_warning(generate_scatterplot(mc_simulation_results, scenario_id = "RS-50"))
+})
+
+test_that("loss_scatterplot() functions", {
+  data(mc_simulation_results)
+  mc_simulation_results %>% filter(scenario_id == "RS-50") %>%
+    unnest(.data$results) -> my_results
+  gg <- loss_scatterplot(my_results)
+  expect_s3_class(gg, "gg")
+})
+
+test_that("Exposure histogram", {
+  data(mc_simulation_results)
+  gg <- mc_simulation_results %>% filter(scenario_id == "RS-50") %>%
+    unnest(.data$results) %>% exposure_histogram()
+  expect_s3_class(gg, "gg")
+})
+
+test_that("Exposure histogram with VaR line", {
+  data(mc_simulation_results)
+  gg <- mc_simulation_results %>% filter(scenario_id == "RS-50") %>%
+    unnest(.data$results) %>% exposure_histogram(show_var_95 = TRUE)
   expect_s3_class(gg, "gg")
 })